add sqs with sns example

+++ src/main/java/poc_aws/poc_tests/sqs/SQSClientOperations.java
@@ -47,7 +47,9 @@ public class SQSClientOperations {
             userChoice = sc.nextLine();
             switch (userChoice){
                 case "deleteQueue":
-                    client.deleteQeueue(queueUrl);
+                    if (queueUrl != null) {
+                        client.deleteQeueue(queueUrl);
+                    }
                     queueUrl = null;
                     break;
                 case "readAll":
